requirements.txt
pymupdf
pytesseract==0.3.10
ocrmypdf>=15.0.0
flask==3.0.0
pymongo==4.6.0
google-generativeai
python-dotenv==1.0.0
pillow
werkzeug==3.0.1
qdrant-client>=1.7.0
